moved ApplicationConfirm to components dir
